moved dd_synthesis.cpp/hpp accordingly, Removed ddsynthesis related functions, moved necessary functions to encoding.hpp and refactored the tests accordingly
