+++ src/CMakeLists.txt
@@ -1,15 +1,18 @@
 add_library(${PROJECT_NAME}
             ${CMAKE_CURRENT_SOURCE_DIR}/algorithms/synthesis/syrec_synthesis.cpp
             ${CMAKE_CURRENT_SOURCE_DIR}/algorithms/simulation/simple_simulation.cpp
+			${CMAKE_CURRENT_SOURCE_DIR}/algorithms/synthesis/dd_synthesis.cpp
+
 
             ${${PROJECT_NAME}_SOURCE_DIR}/include/algorithms/simulation/simple_simulation.hpp
             ${${PROJECT_NAME}_SOURCE_DIR}/include/algorithms/synthesis/syrec_synthesis.hpp
+		    ${${PROJECT_NAME}_SOURCE_DIR}/include/algorithms/synthesis/dd_synthesis.hpp
+		    ${${PROJECT_NAME}_SOURCE_DIR}/include/algorithms/synthesis/encoding.hpp
 
             ${CMAKE_CURRENT_SOURCE_DIR}/core/syrec/parser.cpp
             ${CMAKE_CURRENT_SOURCE_DIR}/core/syrec/program.cpp
             ${CMAKE_CURRENT_SOURCE_DIR}/core/syrec/variable.cpp
             ${CMAKE_CURRENT_SOURCE_DIR}/core/truthTable/truth_table.cpp
-			${CMAKE_CURRENT_SOURCE_DIR}/core/truthTable/dd_synthesis.cpp
 			${CMAKE_CURRENT_SOURCE_DIR}/core/io/pla_parser.cpp
 
             ${${PROJECT_NAME}_SOURCE_DIR}/include/core/gate.hpp
@@ -24,7 +27,6 @@ add_library(${PROJECT_NAME}
             ${${PROJECT_NAME}_SOURCE_DIR}/include/core/syrec/variable.hpp
             ${${PROJECT_NAME}_SOURCE_DIR}/include/core/utils/timer.hpp
             ${${PROJECT_NAME}_SOURCE_DIR}/include/core/truthTable/truth_table.hpp
-			${${PROJECT_NAME}_SOURCE_DIR}/include/core/truthTable/dd_synthesis.hpp
 			${${PROJECT_NAME}_SOURCE_DIR}/include/core/io/pla_parser.hpp)
 
 # set include directories
