+++ test/circuits/CNOT.pla
@@ -1,0 +1,6 @@
+.i 2
+.o 2
+01 01
+10 11
+11 10
+.e
