+++ test/circuits/CUSTOMTT1.pla
@@ -1,0 +1,10 @@
+.i 3
+.o 3
+001 001
+010 100
+011 101
+100 111
+101 010
+110 011
+111 110
+.e
