+++ test/circuits/CUSTOMTT2.pla
@@ -1,0 +1,10 @@
+.i 3
+.o 3
+001 011
+010 001
+011 010
+100 100
+101 111
+110 101
+111 110
+.e
