+++ test/circuits/EXTENDTT.pla
@@ -1,0 +1,5 @@
+.i 2
+.o 2
+-11 111
+1-0 101
+.e
