+++ test/circuits/HUFF2.pla
@@ -1,0 +1,5 @@
+.i 2
+.o 2
+01 11
+1- 11
+.e
