+++ test/circuits/Toffoli.pla
@@ -1,0 +1,10 @@
+.i 3
+.o 3
+001 001
+010 010
+011 011
+100 100
+101 101
+110 111
+111 110
+.e
