+++ test/unittests/test_huffman.cpp
@@ -1,15 +1,15 @@
-#include "core/truthTable/truth_table.hpp"
+#include "algorithms/synthesis/encoding.hpp"
+#include "core/io/pla_parser.hpp"
 
 #include "gtest/gtest.h"
 
 using namespace dd::literals;
 using namespace syrec;
 
-class TruthTableHuff: public testing::TestWithParam<std::string> {
+class TestHuff: public testing::Test {
 protected:
-    dd::QubitCount                 nqubits = 3U;
-    TruthTable                     tt{};
-    std::unique_ptr<dd::Package<>> dd;
+    TruthTable  tt{};
+    std::string test_circuits_dir = "./circuits/";
 
     TruthTable::Cube::Value emptyVal;
 
@@ -22,25 +22,12 @@ protected:
     TruthTable::Cube c001{};
     TruthTable::Cube c010{};
     TruthTable::Cube c011{};
-    TruthTable::Cube c100{};
-    TruthTable::Cube c101{};
-    TruthTable::Cube c110{};
-    TruthTable::Cube c111{};
-
-    TruthTable::Cube c_11{};
-    TruthTable::Cube c1_0{};
 
     TruthTable::Cube c0_{};
-    TruthTable::Cube c1_{};
-
     TruthTable::Cube c0__{};
     TruthTable::Cube c1__{};
 
-    TruthTable::Cube c10_{};
-
     void SetUp() override {
-        dd = std::make_unique<dd::Package<>>(nqubits);
-
         c00 = TruthTable::Cube::fromInteger(0b00U, 2U);
         c01 = TruthTable::Cube::fromInteger(0b01U, 2U);
         c10 = TruthTable::Cube::fromInteger(0b10U, 2U);
@@ -50,18 +37,6 @@ protected:
         c001 = TruthTable::Cube::fromInteger(0b001U, 3U);
         c010 = TruthTable::Cube::fromInteger(0b010U, 3U);
         c011 = TruthTable::Cube::fromInteger(0b011U, 3U);
-        c100 = TruthTable::Cube::fromInteger(0b100U, 3U);
-        c101 = TruthTable::Cube::fromInteger(0b101U, 3U);
-        c110 = TruthTable::Cube::fromInteger(0b110U, 3U);
-        c111 = TruthTable::Cube::fromInteger(0b111U, 3U);
-
-        c_11.emplace_back(emptyVal);
-        c_11.emplace_back(true);
-        c_11.emplace_back(true);
-
-        c1_0.emplace_back(true);
-        c1_0.emplace_back(emptyVal);
-        c1_0.emplace_back(false);
 
         c0_.emplace_back(false);
         c0_.emplace_back(emptyVal);
@@ -70,27 +45,20 @@ protected:
         c0__.emplace_back(emptyVal);
         c0__.emplace_back(emptyVal);
 
-        c1_.emplace_back(true);
-        c1_.emplace_back(emptyVal);
-
         c1__.emplace_back(true);
         c1__.emplace_back(emptyVal);
         c1__.emplace_back(emptyVal);
-
-        c10_.emplace_back(true);
-        c10_.emplace_back(false);
-        c10_.emplace_back(emptyVal);
     }
 };
 
-TEST_F(TruthTableHuff, CUSTOM1) {
-    tt.insert(c01, c01);
-    tt.insert(c10, c01);
-    tt.insert(c11, c11);
+TEST_F(TestHuff, HUFF1) {
+    std::string circHUFF1 = test_circuits_dir + "HUFF1.pla";
+
+    EXPECT_TRUE(read_pla(tt, circHUFF1));
 
     EXPECT_EQ(tt.ioCube().size(), 3U);
 
-    tt.extend();
+    extend(tt);
 
     EXPECT_EQ(tt.ioCube().size(), 4U);
 
@@ -102,7 +70,7 @@ TEST_F(TruthTableHuff, CUSTOM1) {
 
     EXPECT_EQ(tt.nOutputs(), 2U);
 
-    tt.encodeHuffman();
+    encodeHuffman(tt);
 
     EXPECT_EQ(tt.nOutputs(), 2U);
 
@@ -129,14 +97,14 @@ TEST_F(TruthTableHuff, CUSTOM1) {
     EXPECT_EQ(search3->second, c11);
 }
 
-TEST_F(TruthTableHuff, CUSTOM2) {
-    tt.insert(c01, c11);
-    tt.insert(c10, c11);
-    tt.insert(c11, c11);
+TEST_F(TestHuff, HUFF2) {
+    std::string circHUFF2 = test_circuits_dir + "HUFF2.pla";
 
-    EXPECT_EQ(tt.ioCube().size(), 3U);
+    EXPECT_TRUE(read_pla(tt, circHUFF2));
 
-    tt.extend();
+    EXPECT_EQ(tt.ioCube().size(), 2U);
+
+    extend(tt);
 
     EXPECT_EQ(tt.ioCube().size(), 4U);
 
@@ -148,7 +116,7 @@ TEST_F(TruthTableHuff, CUSTOM2) {
 
     EXPECT_EQ(tt.nOutputs(), 2U);
 
-    tt.encodeHuffman();
+    encodeHuffman(tt);
 
     EXPECT_EQ(tt.nOutputs(), 3U);
 
@@ -170,7 +138,7 @@ TEST_F(TruthTableHuff, CUSTOM2) {
 
     EXPECT_EQ(tt.nInputs(), 2U);
 
-    tt.augmentWithConstants();
+    augmentWithConstants(tt);
 
     EXPECT_EQ(tt.nInputs(), 3U);
 
@@ -181,71 +149,4 @@ TEST_F(TruthTableHuff, CUSTOM2) {
 
         EXPECT_TRUE(search4 != tt.ioCube().end());
     }
-}
-
-TEST_F(TruthTableHuff, CUSTOM3) {
-    tt.insert(c000, c010);
-    tt.insert(c001, c010);
-    tt.insert(c010, c100);
-    tt.insert(c011, c100);
-    tt.insert(c100, c011);
-    tt.insert(c101, c010);
-    tt.insert(c110, c010);
-    tt.insert(c111, c001);
-
-    EXPECT_EQ(tt.ioCube().size(), 8U);
-
-    tt.extend();
-
-    EXPECT_EQ(tt.ioCube().size(), 8U);
-
-    auto search1 = tt.ioCube().find(c000);
-
-    EXPECT_TRUE(search1 != tt.ioCube().end());
-
-    EXPECT_EQ(search1->second, c010);
-
-    EXPECT_EQ(tt.nOutputs(), 3U);
-
-    tt.encodeHuffman();
-
-    EXPECT_EQ(tt.nOutputs(), 3U);
-
-    TruthTable::Cube::Vector encInput1{c001, c001, c101, c110};
-
-    for (const auto& in1: encInput1) {
-        auto search2 = tt.ioCube().find(in1);
-
-        EXPECT_TRUE(search2 != tt.ioCube().end());
-
-        EXPECT_EQ(search2->second, c0__);
-    }
-
-    TruthTable::Cube::Vector encInput2{c010, c011};
-
-    for (const auto& in2: encInput2) {
-        auto search3 = tt.ioCube().find(in2);
-
-        EXPECT_TRUE(search3 != tt.ioCube().end());
-
-        EXPECT_EQ(search3->second, c10_);
-    }
-
-    auto search4 = tt.ioCube().find(c100);
-
-    EXPECT_TRUE(search4 != tt.ioCube().end());
-
-    EXPECT_EQ(search4->second, c111);
-
-    auto search5 = tt.ioCube().find(c111);
-
-    EXPECT_TRUE(search5 != tt.ioCube().end());
-
-    EXPECT_EQ(search5->second, c110);
-
-    EXPECT_EQ(tt.nInputs(), 3U);
-
-    tt.augmentWithConstants();
-
-    EXPECT_EQ(tt.nInputs(), 3U);
 }
